# Conflicts:
#	images/downloader/Dockerfile
#	images/onnx-predictor-cpu/Dockerfile
#	images/onnx-predictor-gpu/Dockerfile
#	images/python-predictor-cpu/Dockerfile
#	images/python-predictor-gpu/Dockerfile
#	images/tensorflow-predictor/Dockerfile
#	pkg/cortex/downloader/download.py
#	pkg/cortex/serve/cortex_internal/lib/api/api.py
#	pkg/cortex/serve/cortex_internal/lib/api/predictor.py
#	pkg/cortex/serve/cortex_internal/lib/client/onnx.py
#	pkg/cortex/serve/cortex_internal/lib/client/python.py
#	pkg/cortex/serve/cortex_internal/lib/client/tensorflow.py
#	pkg/cortex/serve/cortex_internal/lib/model/cron.py
#	pkg/cortex/serve/cortex_internal/lib/model/model.py
#	pkg/cortex/serve/cortex_internal/lib/model/tfs.py
#	pkg/cortex/serve/cortex_internal/lib/model/tree.py
#	pkg/cortex/serve/cortex_internal/lib/model/validation.py
#	pkg/cortex/serve/cortex_internal/serve/envsubst.py
#	pkg/cortex/serve/cortex_internal/serve/serve.py
#	pkg/cortex/serve/init/bootloader.sh
#	pkg/cortex/serve/init/script.py
#	pkg/cortex/serve/log_config.yaml
#	pkg/cortex/serve/start/batch.py

+++ pkg/operator/resources/resources.go
@@ -235,7 +235,7 @@ func patchAPI(apiConfig *userconfig.API, configFileName string, force bool) (*sp
 	}
 
 	if deployedResource.Kind != userconfig.TrafficSplitterKind {
-		bytes, err := config.AWS.ReadBytesFromS3(config.Cluster.Bucket, prevAPISpec.ProjectKey)
+		bytes, err := config.ReadBytesFromBucket(prevAPISpec.ProjectKey)
 		if err != nil {
 			return nil, "", err
 		}
@@ -293,10 +293,16 @@ func DeleteAPI(apiName string, keepCache bool) (*schema.DeleteResponse, error) {
 					return realtimeapi.DeleteAPI(apiName, keepCache)
 				},
 				func() error {
-					return batchapi.DeleteAPI(apiName, keepCache)
+					if config.Provider == types.AWSProviderType {
+						return batchapi.DeleteAPI(apiName, keepCache)
+					}
+					return nil
 				},
 				func() error {
-					return trafficsplitter.DeleteAPI(apiName, keepCache)
+					if config.Provider == types.AWSProviderType {
+						return trafficsplitter.DeleteAPI(apiName, keepCache)
+					}
+					return nil
 				},
 			)
 			if err != nil {
